+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [0.0.6]
+
+### Features
+
+-   HTTP proxy support ([#38](https://github.com/Scalr/scalr-vscode/issues/38))
+-   Scalr Workspaces & Runs Commands ([#36]https://github.com/Scalr/scalr-vscode/issues/36))
+
 ## [0.0.5]
 
 ### Features
@@ -44,7 +51,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ---
 
-[Unreleased]: https://github.com/Scalr/scalr-vscode/compare/v0.0.5...HEAD
+[Unreleased]: https://github.com/Scalr/scalr-vscode/compare/v0.0.6...HEAD
+[0.0.6]: https://github.com/Scalr/scalr-vscode/releases/tag/v0.0.6
 [0.0.5]: https://github.com/Scalr/scalr-vscode/releases/tag/v0.0.5
 [0.0.4]: https://github.com/Scalr/scalr-vscode/releases/tag/v0.0.4
 [0.0.3]: https://github.com/Scalr/scalr-vscode/releases/tag/v0.0.3